Add PySpark code mimicking SearchVariantsExampleBRCA1
The idea is we create an RDD in Scala, transform that into an RDD of
JSON String, and save that to a text file. From PySpark, we load that
text file and decode the JSON entries, then perform main tasks (such as
filtering, counting) with Python code (lambda in the example).